* master: (525 commits)
  BGE Action Actuator fixes:   * Ping pong now has the same behavior as 2.59.   * Setting the frame on a animation playing in reverse now works correctly   * The frames for continuous were a frame or so behind, this is now fixed.
  fix for crash when entering in non unicode ascii chars.
  utf8 osx - up and down keys are producing utf8 output. muting them out in cocoa
  added function BKE_library_filepath_set which sync's the libraries absolute path when setting to a relative value, before this you could never be sure if a libraries absolute path was valid or not because the user might have changed the relative library path in the outliner, now setting from the outliner and py/rna syncs the absolute path.
  Another set of UI messages fixes and tweaks! No functional changes.
  Fix for commit 41227 (Some opening comment tags (/*) were lost!).
  Remove some more $Id$ that still were left after r41227 and r41228.
  Remove the svn:keywords property.
  fix for error in last commit and remove some more Id's I missed
  remove $Id: tags after discussion on the mailign list: http://markmail.org/message/fp7ozcywxum3ar7n
  Typo fixes for the typo fixes. :D
  fix some typo's
  - fix for error with utf8 textinput for buttons - ensure input is valid utf8 from ghost and NULL then complain if its not. - added function to get utf8 size BLI_str_utf8_size()
  Code refactoring: split main 3d view drawing function into object drawing and info overlay drawing functions.
  RNA: fix collection iterator issue in c++ api with msvc.
  committing 'a' to merge into the tag
  minor formatting edits (80 char width)
  set pythons 'sys.stdout' encoding to utf-8 and use surrogateescape error handler.
  rna access to event.unicode so scripts can get unicode text input.
  py/rna api crash fix: getting event.ascii would crash blender (strange nobody noticed this, seems its been there for over a year???).
  ...

Conflicts:
	source/blender/editors/space_outliner/outliner_draw.c
	source/blender/python/mathutils/mathutils_Color.c
	source/blender/python/mathutils/mathutils_Color.h
	source/blender/python/mathutils/mathutils_Euler.c
	source/blender/python/mathutils/mathutils_Euler.h
	source/blender/python/mathutils/mathutils_Matrix.c
	source/blender/python/mathutils/mathutils_Matrix.h
	source/blender/python/mathutils/mathutils_Vector.c
	source/blender/python/mathutils/mathutils_Vector.h

+++ source/blender/editors/include/UI_interface_icons.h
@@ -1,6 +1,4 @@
 /*
- * $Id$
- *
  * ***** BEGIN GPL LICENSE BLOCK *****
  *
  * This program is free software; you can redistribute it and/or
@@ -50,8 +48,8 @@ typedef struct IconFile {
 #define ICON_DEFAULT_HEIGHT 16
 #define ICON_DEFAULT_WIDTH	16
 
-#define ICON_DEFAULT_HEIGHT_SCALE	(UI_UNIT_Y * 0.8f)
-#define ICON_DEFAULT_WIDTH_SCALE	(UI_UNIT_X * 0.8f)
+#define ICON_DEFAULT_HEIGHT_SCALE ((int)(UI_UNIT_Y * 0.8f))
+#define ICON_DEFAULT_WIDTH_SCALE  ((int)(UI_UNIT_X * 0.8f))
 
 #define PREVIEW_DEFAULT_HEIGHT 96
